fix: harmonize response across single and cluster
for the cluster the response perviously has been an array of each
individual nodes responses. Now, return just the first one.

Maybe this is not the best thing to do and we should return each nodes
individual response, or throw if they differ. But I think its fine for
now, we can iterate as needed.